implemented GSTIN, TAN, IGST(18%)

+++ upgrades/src/main/java/com/boost/upgrades/data/api_model/PurchaseOrder/request/PaymentDetails.kt
@@ -5,5 +5,5 @@ data class PaymentDetails(
     val Discount: Int,
     val PaymentChannelProvider: String,
     val TaxDetails: TaxDetails,
-    val TotalPrice: Int
+    val TotalPrice: Double
 )
