+++ upgrades/src/main/java/com/boost/upgrades/data/api_model/PurchaseOrder/request/Widget.kt
@@ -9,7 +9,7 @@ data class Widget(
     val IsCancellable: Boolean,
     val IsRecurringPayment: Boolean,
     val Name: String,
-    val Price: Int,
+    val Price: Double,
     val RecurringPaymentFrequency: String,
     val WidgetKey: String,
     val Quantity: Int
