+++ upgrades/src/main/java/com/boost/upgrades/data/api_model/PurchaseOrder/response/Result.kt
@@ -2,6 +2,6 @@ package com.boost.upgrades.data.api_model.PurchaseOrder.response
 
 data class Result(
     val OrderId: String,
-    val TotalPrice: Int,
+    val TotalPrice: Double,
     val TransactionId: String
 )
